Rename screenshot module into components module.
