+++ app/build.gradle
@@ -64,8 +64,7 @@ android {
 }
 
 dependencies {
-    implementation project(":screenshots")
-
+    implementation project(':components')
     //Desugaring for use of java.time in api lower then 26
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:2.0.2'
