+++ app/src/androidTest/java/com/appunite/loudius/LoginScreenTest.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.appunite.loudius.screenshots.theme.LoudiusTheme
+import com.appunite.loudius.components.theme.LoudiusTheme
 import com.appunite.loudius.ui.login.LoginScreen
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
