+++ app/src/androidTest/java/com/appunite/loudius/PullRequestsScreenTest.kt
@@ -20,8 +20,8 @@ import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.appunite.loudius.screenshots.components.countingResource
-import com.appunite.loudius.screenshots.theme.LoudiusTheme
+import com.appunite.loudius.components.components.countingResource
+import com.appunite.loudius.components.theme.LoudiusTheme
 import com.appunite.loudius.ui.pullrequests.PullRequestsScreen
 import com.appunite.loudius.util.IdlingResourceExtensions.toIdlingResource
 import com.appunite.loudius.util.MockWebServerRule
