+++ app/src/main/java/com/appunite/loudius/ui/pullrequests/PullRequestsScreen.kt
@@ -37,18 +37,18 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.appunite.loudius.R
 import com.appunite.loudius.common.Constants
+import com.appunite.loudius.components.components.LoudiusFullScreenError
+import com.appunite.loudius.components.components.LoudiusListIcon
+import com.appunite.loudius.components.components.LoudiusListItem
+import com.appunite.loudius.components.components.LoudiusLoadingIndicator
+import com.appunite.loudius.components.components.LoudiusPlaceholderText
+import com.appunite.loudius.components.components.LoudiusText
+import com.appunite.loudius.components.components.LoudiusTextStyle
+import com.appunite.loudius.components.components.LoudiusTopAppBar
+import com.appunite.loudius.components.theme.LoudiusTheme
 import com.appunite.loudius.network.model.PullRequest
-import com.appunite.loudius.screenshots.components.LoudiusFullScreenError
-import com.appunite.loudius.screenshots.components.LoudiusListIcon
-import com.appunite.loudius.screenshots.components.LoudiusListItem
-import com.appunite.loudius.screenshots.components.LoudiusLoadingIndicator
-import com.appunite.loudius.screenshots.components.LoudiusPlaceholderText
-import com.appunite.loudius.screenshots.components.LoudiusText
-import com.appunite.loudius.screenshots.components.LoudiusTextStyle
-import com.appunite.loudius.screenshots.components.LoudiusTopAppBar
-import com.appunite.loudius.screenshots.theme.LoudiusTheme
 import java.time.LocalDateTime
-import com.appunite.loudius.screenshots.R as screenshotsR
+import com.appunite.loudius.components.R as componentsR
 
 typealias NavigateToReviewers = (String, String, String, String) -> Unit
 
@@ -189,7 +189,7 @@ private fun RepoDetails(modifier: Modifier, pullRequestTitle: String, repository
 private fun EmptyListPlaceholder(padding: PaddingValues) {
     Box(modifier = Modifier.padding(padding)) {
         LoudiusPlaceholderText(
-            textId = screenshotsR.string.you_dont_have_any_pull_request,
+            textId = componentsR.string.you_dont_have_any_pull_request,
         )
     }
 }
