+++ components/src/main/java/com/appunite/loudius/components/theme/Color.kt
@@ -14,19 +14,21 @@
  * limitations under the License.
  */
 
-package com.appunite.loudius.util
+package com.appunite.loudius.components.theme
 
-import androidx.compose.ui.test.IdlingResource
-import com.appunite.loudius.components.common.CountingIdlingResource
+import androidx.compose.ui.graphics.Color
 
-object IdlingResourceExtensions {
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
 
-    fun CountingIdlingResource.toIdlingResource(): IdlingResource = object :
-        IdlingResource {
-        override val isIdleNow: Boolean
-            get() = this@toIdlingResource.isIdleNow
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
 
-        override fun getDiagnosticMessageIfBusy(): String =
-            this@toIdlingResource.getDiagnosticMessageIfBusy()
-    }
-}
+val White99 = Color(0xFFFFFBFE)
+val Black90 = Color(0xFF1C1B1F)
+
+val NeutralVariant30 = Color(0xFF49454F)
+val NeutralVariant80 = Color(0xFFCAC4D0)
+val Error20 = Color(0xFF601410)
