+++ components/src/main/res/drawable/person_outline_24px.xml
