+++ components/src/main/res/raw/loading_indicator.json
