+++ components/src/main/res/values/strings.xml
