--- screenshots/src/main/java/com/appunite/loudius/screenshots/common/CountingIdlingResource.kt
@@ -1,33 +1,0 @@
-/*
- * Copyright 2023 AppUnite S.A.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.appunite.loudius.screenshots.common
-
-import java.util.concurrent.atomic.AtomicInteger
-
-class CountingIdlingResource(val name: String) {
-    private var counter = AtomicInteger(0)
-    val isIdleNow: Boolean get() = counter.get() == 0
-    fun getDiagnosticMessageIfBusy(): String = "$name is busy"
-
-    fun increment() {
-        counter.incrementAndGet()
-    }
-
-    fun decrement() {
-        counter.decrementAndGet()
-    }
-}
