--- screenshots/src/main/java/com/appunite/loudius/screenshots/components/IdlingResourceWrapper.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright 2023 AppUnite S.A.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.appunite.loudius.screenshots.components
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import com.appunite.loudius.screenshots.common.CountingIdlingResource
-
-val countingResource = CountingIdlingResource("IdlingResourceWrapper")
-
-@Composable
-fun IdlingResourceWrapper(content: @Composable () -> Unit) {
-    DisposableEffect(Unit) {
-        countingResource.increment()
-        onDispose {
-            countingResource.decrement()
-        }
-    }
-    content()
-}
