--- screenshots/src/main/java/com/appunite/loudius/screenshots/components/LoudiusPlaceholderText.kt
@@ -1,53 +1,0 @@
-/*
- * Copyright 2023 AppUnite S.A.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.appunite.loudius.screenshots.components
-
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.appunite.loudius.screenshots.R
-import com.appunite.loudius.screenshots.theme.LoudiusTheme
-
-@Composable
-fun LoudiusPlaceholderText(@StringRes textId: Int) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        contentAlignment = Alignment.Center,
-    ) {
-        LoudiusText(
-            text = stringResource(id = textId),
-            style = LoudiusTextStyle.ScreenContent,
-        )
-    }
-}
-
-@Preview(showSystemUi = true)
-@Composable
-fun PreviewLoudiusPlaceholderText() {
-    LoudiusTheme {
-        LoudiusPlaceholderText(R.string.you_dont_have_any_pull_request)
-    }
-}
