+++ settings.gradle
@@ -15,4 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "Loudius"
 include ':app'
 include ':custom-ktlint-rules'
-include ':screenshots'
+include ':components'
